Support non-standard npm aliases
Unlike node, npm doesn't have great support for transparently running
multiple versions in parallel. Instead, a common way to run npm3
and npm2 in parallel is to use one to globally install the other,
with a separate alias (https://www.npmjs.com/package/npm3).

The boilerplate currently explicitly uses the binary named "npm" to
run scripts, rather than using the npm binary that was used to
invoke the script. So "npm3 run npmcheckversion" can fail with
a complaint that you need to be using npm v3.

This change uses the npm_execpath environment variable to run
npm scripts with whichever npm binary was used to invoke the script.