Use semantic structure and replace div with header element
For semantic structure, <header> tag is missing in the code and <footer> tag is already in the code. And we can see that the code is broken it into header component but not using <header> tag.

So, replacing <header> with <div> and removing caps from <a> & <img> tag for better semantic structure.